// lib/api/generated/models/flagKind.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * CKS Slavia API
 * API panelu klubowego CKS Slavia Ruda Śląska
 * OpenAPI spec version: 2.2.0.1+1
 */

export type FlagKind = typeof FlagKind[keyof typeof FlagKind];


export const FlagKind = {
  stable: 'stable',
  experimental: 'experimental',
} as const;