// lib/api/generated/models/siteStats.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * CKS Slavia API
 * API panelu klubowego CKS Slavia Ruda Śląska
 * OpenAPI spec version: 1.1.0.17+18
 */

export interface SiteStats {
  /** @minimum 0 */
  active_users: number;
  /** @minimum 0 */
  athlete_profiles: number;
  /** @minimum 0 */
  cms_pages: number;
  /** @minimum 0 */
  cms_published: number;
  /** @minimum 0 */
  feature_flags: number;
  /** @minimum 0 */
  results_pending: number;
  /** @minimum 0 */
  results_total: number;
  /** @minimum 0 */
  system_logs: number;
  /** @minimum 0 */
  users: number;
}